+++ alexber/utils/processinvokes.py
@@ -3,6 +3,7 @@ import os as _os
 import subprocess
 import threading
 from concurrent.futures import ThreadPoolExecutor
+from . importer import importer
 
 executor = None
 default_log_name  = None
@@ -15,20 +16,9 @@ default_log_subprocess_cls = None
 # that is itself adoptation of https://gist.github.com/bgreenlee/1402841
 
 
-class LogPipe(object):
+class BasePipe(object):
     def __init__(self, *args, **kwargs):
-        """
-        Setup the logger with logName and loglevel
-        Create read and write file descriptor of the pipe
-        """
-        logName = kwargs.pop('logName', default_log_name)
-        logLevel = kwargs.pop('logLevel', default_log_level)
-
-        super().__init__(**kwargs)
-
-        self.logger = logging.getLogger(logName)
-
-        self.level = logLevel
+        super().__init__(*args, **kwargs)
         self.fdRead, self.fdWrite = _os.pipe()
         self.fdWriteLock = threading.RLock()
 
@@ -36,25 +26,41 @@ class LogPipe(object):
             self.is_closed = False
 
     def fileno(self):
-        """Return the write file descriptor of the pipe
+        """
+        Return the write file descriptor of the pipe
         """
         return self.fdWrite
 
+    def processLine(self, line):
+        """
+        Hook to be overridden in the sub-class.
+        """
+        pass
+
     def run(self):
-        """Run the thread, logging everything.
+        """
+        Run the thread, process output line by line.
         """
         with _os.fdopen(self.fdRead) as pipeReader:
             for line in iter(pipeReader.readline, ''):
-                self.logger.log(self.level, line.rstrip('\r\n'))
+                self.processLine(line)
 
     def breakPipe(self):
         self.close()
 
+    def cleanUp(self):
+        """
+        Hook to be overridden in the sub-class.
+        """
+        pass
+
     def close(self):
-        """Close the write end of the pipe.
+        """
+        Close the write end of the pipe.
         """
         with self.fdWriteLock:
             if not self.is_closed:
+                self.cleanUp()
                 _os.close(self.fdWrite)
             self.is_closed = True
 
@@ -68,39 +74,85 @@ class LogPipe(object):
         self.close()
 
 
+class LogPipe(BasePipe):
+    def __init__(self, *args, **kwargs):
+        """
+        Setup the logger with logName and loglevel
+        Create read and write file descriptor of the pipe
+        """
+        logName = kwargs.pop('logName', default_log_name)
+        logLevel = kwargs.pop('logLevel', default_log_level)
+
+        super().__init__(*args, **kwargs)
+
+        self.logger = logging.getLogger(logName)
+
+        self.level = logLevel
+
+    def processLine(self, line):
+        self.logger.log(self.level, line.rstrip('\r\n'))
+
+class FilePipe(BasePipe):
+    def __init__(self, *args, **kwargs):
+        filename = kwargs.pop('fileName', None)
+        if filename is None:
+            raise ValueError('fileName should be specified')
+
+        super().__init__(*args, **kwargs)
+
+        d = {
+            'file': filename,
+            'mode': 'w',
+            **kwargs
+        }
+
+        self.f = open(**d)
+
+    def processLine(self, line):
+        aline = f"{line.rstrip()}\n"
+        self.f.write(aline)
+
+    def cleanUp(self):
+        if self.f is not None:
+            self.f.close()
 
 
-class LoggigSubProcessCall(object):
+
+
+
+class LogSubProcessCall(object):
     def __init__(self, *args, **kwargs):
 
-        self.logpipe = kwargs.pop('pipe')
+        self.pipe = kwargs.pop('pipe')
         d = kwargs.pop('popen', {})
 
         self.popenargs = d.pop('args', [])
         self.popenkwargs = d.pop('kwargs', {})
-        super().__init__(**kwargs)
+        super().__init__(*args, **kwargs)
 
     def calc_subprocess_run_kwargs(self):
-        kwargs= {'stdout':self.logpipe, 'stderr':subprocess.STDOUT,
+        kwargs= {'stdout':self.pipe, 'stderr':subprocess.STDOUT,
             'text':True, 'bufsize':1, 'check':True,
-            **self.popenkwargs}
+                 **self.popenkwargs}
         return kwargs
 
     def run_sub_process(self):
-        f = executor.submit(self.logpipe.run)
+        f = executor.submit(self.pipe.run)
         try:
             kwargs = self.calc_subprocess_run_kwargs()
             process = subprocess.run(self.popenargs, **kwargs)
         finally:
-            self.logpipe.breakPipe()
+            self.pipe.breakPipe()
             f.result()
-
 
 
 
 def run_sub_process(*args, **kwargs):
     """
     This method run subprocess and logs it's out to the logger.
+    This method is sophisticated decorator to subprocess.run(). It is useful, when your subprocess
+    run's a lot of time and you're interesting to receive it's stdout and stderr. By default, it's streamed to log.
+    You can easily customize this behavior, see `initConig()` method.
 
     This method is sophisticated decorator to subprocess.run(). See it's docstring for more information. Note, that
     some parameters (cwd, for example) that can be used in popenkwargs are listed in Popen constructor.
@@ -108,6 +160,7 @@ def run_sub_process(*args, **kwargs):
     logPipe is customizable object that essentially forwards output from subprocess to the logger using
     logName and logLevel (see below).
 
+    default_log_subprocess_cls by default is LogSubProcessCall.
     default_logpipe_cls by default is LogPipe.
     logName by default is processinvokes.
     logLevel by default is logger.INFO.
@@ -129,21 +182,18 @@ def run_sub_process(*args, **kwargs):
           Essentially, no OS-level buffering between process, provided that call to write contains a newline character.
         If the exit code of subprocess is non-zero, it raises a CalledProcessError.
 
-
-
-
     It is generally not advice to override them, but you can if you know, what you're doing.
-
-
-
 
     :param args: will be passed as popenargs to subprocess.run() method
     :param roughly kwargs['kwargs'] will be passed as popenkwargs to subprocess.run() method
     :param roughly kwargs['logPipe']['cls'] or default_logpipe_cls (if first one is empty)
                                              will be passed as logPipeCls to create logPipe.
+                                             Can be class or str.
     :param roughly kwargs['logPipe']['kwargs'] will be passed as kwargs to logPipeCls to create logPipe.
     :param roughly kwargs['logSubprocess']['cls'] or default_log_subprocess_cls (if first one is empty)
-                                             will be passed as logSubProcessCls to create LoggigSubProcessCall.
+                                             will be passed as logSubProcessCls to create LogSubProcessCall.
+                                              Can be class or str.
+    :param roughly kwargs['logSubprocess']['kwargs'] will be passed as kwargs to logSubprocess.
 
     :return:
     """
@@ -152,17 +202,28 @@ def run_sub_process(*args, **kwargs):
 
     logPipe_p = kwargs.pop('logPipe', {})
     logPipeCls = logPipe_p.pop('cls', default_logpipe_cls)
+    if isinstance(logPipeCls, str):
+        logPipeCls = importer(logPipeCls)
+
     logPipeKwargs = logPipe_p.pop('kwargs', {})
     callKwargs = kwargs.pop('kwargs', {})
     logSubprocess_p = kwargs.pop('logSubprocess', {})
-    logSubProcessCls = logSubprocess_p.pop('logSubprocess', default_log_subprocess_cls)
+    logSubProcessCls = logSubprocess_p.pop('cls', default_log_subprocess_cls)
+    if isinstance(logSubProcessCls, str):
+        logSubProcessCls = importer(logSubProcessCls)
+    logSubprocessKwargs = logSubprocess_p.pop('kwargs', {})
 
     with logPipeCls(**logPipeKwargs) as logPipe:
-        call = logSubProcessCls(pipe=logPipe,
-                                    **{'popen':
-                                           {'args':args,
-                                           'kwargs':callKwargs}}
-                                    )
+
+        kwargs = {'pipe': logPipe,
+                 'popen':
+                      {'args': args,
+                       'kwargs': callKwargs
+                      },
+                 **logSubprocessKwargs
+                  }
+
+        call = logSubProcessCls(**kwargs)
         call.run_sub_process()
 
 
@@ -173,14 +234,15 @@ def initConfig(**kwargs):
     It is indented to be called in the MainThread.
     This method can be call with empty params.
 
-    :param default_log_name: Optional.
+    :param default_log_name: Optional. - name of the logger where the messages will be streamed to.
                 Default values is: processinvokes
-    :param default_log_level: Optional.
+    :param default_log_level: Optional. - log level to be used in logger.
                 Default values is: logging.INFO
-    :param default_logpipe_cls: Optional.
+    :param default_logpipe_cls: can be class or str. Optional. You can use your custom class for the logging.
+                For example, FilePipe.
                 Default values is: LogPipe
-    :param default_log_subprocess_cls: Optional.
-                Default values is: LoggigSubProcessCall
+    :param default_log_subprocess_cls: can be class or str. Optional.
+                Default values is: LogSubProcessCall
     :param executor: internally used to run sub-process
                 Default values are
                       'max_workers':1,
@@ -207,12 +269,16 @@ def initConfig(**kwargs):
     default_logpipe_cls_p = kwargs.get('default_logpipe_cls', None)
     if default_logpipe_cls_p is None:
         default_logpipe_cls_p = LogPipe
+    elif isinstance(default_logpipe_cls_p, str):
+        default_logpipe_cls_p = importer(default_logpipe_cls_p)
     global default_logpipe_cls
     default_logpipe_cls = default_logpipe_cls_p
 
     default_log_subprocess_cls_p = kwargs.get('default_log_subprocess_cls', None)
     if default_log_subprocess_cls_p is None:
-        default_log_subprocess_cls_p = LoggigSubProcessCall
+        default_log_subprocess_cls_p = LogSubProcessCall
+    elif isinstance(default_log_subprocess_cls_p, str):
+        default_log_subprocess_cls_p = importer(default_log_subprocess_cls_p)
     global default_log_subprocess_cls
     default_log_subprocess_cls = default_log_subprocess_cls_p