+++ alexber/utils/ymlparsers.py
@@ -2,7 +2,7 @@
 This is essentially wrapper arround HiYaPyCo project with streamlined and extended API and couple of work-arrounds.
 
 This module depends on some 3-rd party dependencies, in order to use it, you should have installed first. To do it
-run `pip3 install alex-ber-utils[yml]`.
+run `python3 -m pip install alex-ber-utils[yml]`.
 
 This module doesn't use any package-level variables in hiYaPyCo module, including hiYaPyCo.jinja2env.
 This module do use Jinja2's `Environment`.
@@ -22,7 +22,7 @@ except ImportError:
 
     warning = (
         "You appear to be missing some optional dependencies (jinja2);"
-        "please 'pip3 install alex-ber-utils[yml]'."
+        "please 'python3 -m pip install alex-ber-utils[yml]'."
     )
     warnings.warn(warning, ImportWarning)
     raise
@@ -34,7 +34,7 @@ except ImportError:
 
     warning = (
         "You appear to be missing some optional dependencies (hiyapyco);"
-        "please 'pip3 install alex-ber-utils[yml]'."
+        "please 'python3 -m pip install alex-ber-utils[yml]'."
     )
     warnings.warn(warning, ImportWarning)
     raise
@@ -46,7 +46,7 @@ except AttributeError as e:
 
     warning = (
         "You appear to be missing some optional yml parsing dependencies (hiyapyco should be at least 0.4.16);"
-        "please 'pip3 install alex-ber-utils[yml]'."
+        "please 'python3 -m pip install alex-ber-utils[yml]'."
     )
     warnings.warn(warning, ImportWarning)
     raise ImportError(str(e)) from e
@@ -65,8 +65,9 @@ _safe_dump_d = None
 _load_d = None
 
 from platform import uname as _uname
+from . _ymlparsers_extra import format_template as _format_template
 
-
+format_template = _format_template
 
 class HiYaPyCo(_HiYaPyCo):
     jinja2ctx = None