+++ tests/utils/processinvokes_test.py
@@ -1,8 +1,11 @@
 import logging
 import shlex
 import os as _os
+import tempfile
+from pathlib import Path
 import pytest
 import alexber.utils.processinvokes as processinvokes
+from alexber.utils.processinvokes import LogPipe, LogSubProcessCall
 
 logger = logging.getLogger(__name__)
 process_invokes_logger = None
@@ -12,6 +15,17 @@ _process_invokes_logger_log = None
 def mock_log(mocker):
     ret_mock = mocker.patch.object(process_invokes_logger, 'log', side_effect=_process_invokes_logger_log, autospec=True, spec_set=True)
     return ret_mock
+
+
+@pytest.fixture
+def mock_file(mocker):
+    open_mock = mocker.patch('.'.join(['alexber.utils.processinvokes', 'open']), create=True)
+    mock_close = mocker.MagicMock()
+    mock_write = mocker.MagicMock()
+    open_mock.return_value.close = mock_close
+    open_mock.return_value.write = mock_write
+
+    return open_mock
 
 
 def _reset_processinvokes():
@@ -54,6 +68,77 @@ def test_process_invokes(request, mocker, processinvokesFixture, mock_log):
     pytest.assume(mock_log.call_count == 1)
     (_,logmsg), _ = mock_log.call_args
     pytest.assume(exp_log_msg == logmsg)
+
+class MyLogPipe(LogPipe):
+    pass
+
+class MyLogSubProcessCall(LogSubProcessCall):
+    pass
+
+def test_init_config(request, mocker, processinvokesFixture):
+    logger.info(f'{request._pyfuncitem.name}()')
+
+    logPipeClassName = '.'.join([__name__, MyLogPipe.__name__])
+    logSubProcessCallClassName = '.'.join([__name__, MyLogSubProcessCall.__name__])
+
+    processinvokes.initConfig(**{'default_log_name': 'process_invoke_run',
+                                 'default_logpipe_cls': logPipeClassName,
+                                 'default_log_subprocess_cls': logSubProcessCallClassName,
+                                 })
+
+    pytest.assume(processinvokes.default_logpipe_cls == MyLogPipe)
+    pytest.assume(processinvokes.default_log_subprocess_cls == MyLogSubProcessCall)
+
+
+def test_process_invokes_file_pipe1(request, mocker, processinvokesFixture, mock_file):
+    logger.info(f'{request._pyfuncitem.name}()')
+    exp_log_msg = "simulating run_sub_process"
+    process_invoke_run = f"echo '{exp_log_msg}'"
+    cmd = shlex.split(process_invoke_run)
+
+    process_invoke_cwd = _os.getcwd()
+    filename = "my.log"
+    mock_close = mocker.MagicMock()
+    mock_write = mocker.MagicMock()
+    mock_file.close = mock_close
+    mock_file.write = mock_write
+    processinvokes.initConfig(**{'default_log_name': 'process_invoke_run',
+                                 'default_logpipe_cls': 'alexber.utils.processinvokes.FilePipe'
+                                 })
+
+    processinvokes.run_sub_process(*cmd, **{'cwd':process_invoke_cwd,
+                                                      'logPipe': {
+                                                          'kwargs' : {'fileName': filename}
+                                                      }
+                                                      })
+
+    pytest.assume(mock_file.return_value.close.call_count == 1)
+    pytest.assume(mock_file.return_value.write.call_count > 0)
+    (logmsg,), _ = mock_file.return_value.write.call_args
+    pytest.assume(f'{exp_log_msg}\n' == logmsg)
+
+def test_process_invokes_file_pipe2(request, mocker, processinvokesFixture, mock_file):
+    logger.info(f'{request._pyfuncitem.name}()')
+    exp_log_msg = "simulating run_sub_process"
+    process_invoke_run = f"echo '{exp_log_msg}'"
+    cmd = shlex.split(process_invoke_run)
+
+    process_invoke_cwd = _os.getcwd()
+    filename = "my.log"
+
+
+    processinvokes.run_sub_process(*cmd, **{'cwd':process_invoke_cwd,
+                                                  'logPipe': {
+                                                      'cls': 'alexber.utils.processinvokes.FilePipe',
+                                                      'kwargs' : {'fileName': filename}
+                                                  }
+                                                  })
+
+    pytest.assume(mock_file.return_value.close.call_count == 1)
+    pytest.assume(mock_file.return_value.write.call_count > 0)
+    (logmsg,), _ = mock_file.return_value.write.call_args
+    pytest.assume(f'{exp_log_msg}\n' == logmsg)
+
 
 
 if __name__ == "__main__":